# openapi.yaml
openapi: 3.1.0
info:
  title: SecureFlow API
  description: |
    SecureFlow is a GitHub App that provides automated security scanning for pull requests.
    This API enables programmatic access to findings, repositories, SBOM scanning, and more.
  version: 1.0.0
  contact:
    name: SecureFlow Team
    url: https://github.com/GauravKarakoti/SecureFlow
  license:
    name: MIT
    url: https://github.com/GauravKarakoti/SecureFlow/blob/main/LICENSE
servers:
  - url: https://secure-flow-six.vercel.app
    description: Production server
  - url: http://localhost:9002
    description: Local development server
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth.js JWT session cookie. Obtained via GitHub OAuth flow.
        Sessions have a 1-year maxAge. The session contains user ID, codename, and roles.
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Machine-readable error code
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - TOO_MANY_REQUESTS
            - INTERNAL_SERVER_ERROR
            - DATABASE_ERROR
        message:
          type: string
          description: Human-readable error message (sensitive data scrubbed)
        success:
          type: boolean
          enum: [false]
...
paths:
...
  /api/webhooks/github:
    post:
      summary: GitHub webhook receiver
      description: |
        Receives GitHub App webhook events. Handles pull_request, installation,
        and installation_repositories events. Runs security scans on PRs.
      operationId: handleGitHubWebhook
      tags:
        - Webhooks
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
            enum: [pull_request, installation, installation_repositories]
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
            description: HMAC-SHA256 signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook payload (varies by event type)
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  decision:
                    type: string
                    enum: [PASS, REVIEW REQUIRED, BLOCKED]
                  findingCount:
                    type: integer
                  error:
                    type: string
                  fallback:
                    type: boolean
        "202":
          description: Duplicate delivery or rate limited
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
        "400":
          description: Invalid payload or signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded (50 req/60s)
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Webhook processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...